refactor to move constants, statics, enums to separate file from types.rs, handlers.rs
