+++ src/rust/src/xds/mod.rs
@@ -21,7 +21,8 @@ use crate::bindings::*;
 use crate::ctorust::FromCType;
 use crate::libccxr_exports::time::apply_timing_info;
 use crate::libccxr_exports::time::generate_timing_context;
-use crate::xds::handlers::{do_end_of_xds, TS_START_OF_XDS};
+use crate::xds::constants::TS_START_OF_XDS;
+use crate::xds::handlers::do_end_of_xds;
 use crate::xds::types::{copy_xds_context_from_rust_to_c, CcxDecodersXdsContext};
 use std::os::raw::c_int;
 use std::sync::atomic::Ordering;
